backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
python-dotenv==1.0.1
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.29.0
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
apscheduler==3.10.4
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
rapidfuzz==3.10.0
python-dateutil==2.9.0.post0
tenacity==9.0.0
PyPDF2==3.0.1